requirement.txt
xhtml2pdf
Django==3.0.5
django-widget-tweaks==1.4.8
sqlparse==0.3.1
django-import-export
python-barcode